Add simple code to compare straight c and c++ On 43oh.com there has been a discussion about how to use the usci serial port. I took the posted code and wrote a fabooh version of that code. It turns out that fabooh on at least msp430-gcc is smaller.

+++ fabooh/common.mk
@@ -21,7 +21,6 @@ include $(FABOOH_DIR)/mkfiles/include-msp430g2553in20.mk
 # or by putting it in their makefiles
 FLAGS?=
 USEROBJS?=
-USERCLEAN?=
 LDLIBS?=
 
 # tools we use
@@ -68,7 +67,9 @@ clean:
 	@rm -f $(TARGET)_asm_count.txt
 	@rm -f $(TARGET).hex
 	@rm -f $(TARGET).map
+ifdef USERCLEAN
 	rm -f $(USERCLEAN)
+endif
  
 install: all
 	$(MSPDEBUG) rf2500 "prog $(TARGET).elf"
